+++ src/main/java/org/usfirst/frc/team1294/robot/commands/DeliverGearCommand.java
@@ -4,6 +4,7 @@ import edu.wpi.first.wpilibj.PIDController;
 import edu.wpi.first.wpilibj.PIDSource;
 import edu.wpi.first.wpilibj.PIDSourceType;
 import edu.wpi.first.wpilibj.command.Command;
+import edu.wpi.first.wpilibj.command.CommandGroup;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import org.usfirst.frc.team1294.robot.Robot;
 import org.usfirst.frc.team1294.robot.util.SimplePIDSource;
@@ -14,102 +15,109 @@ import org.usfirst.frc.team1294.robot.util.SimplePIDSource;
  * lift squared up with the gear side of the robot. Uses vision system to move the robot
  * side to side to keep the lift centered. Will time out after 10 seconds.
  */
-public class DeliverGearCommand extends Command {
-  private static final double ANGLE_TOLERANCE_DEGREES = 2.0f;
-  private static final double ANGLE_KP = 0.3f;
-  private static final double ANGLE_KI = 0;
-  private static final double ANGLE_KD = 0;
-
-  private static final double STRAFE_TOLERANCE = 0.5f;
-  private static final double STRAFE_KP = 0.3f;
-  private static final double STRAFE_KI = 0;
-  private static final double STRAFE_KD = 0;
-
-  private static final double APPROACH_TOLERANCE = 0.5f;
-  private static final double APPROACH_KP = 0.3f;
-  private static final double APPROACH_KI = 0;
-  private static final double APPROACH_KD = 0;
-
-  //private final PIDController anglePid;
-  private final PIDController strafePid;
-  //private final PIDController approachPid;
-
-  private double commandedTurnRate = 0;
-  private double commandedStrafeRate = 0;
-  private double commandedApproachRate = 0;
+public class DeliverGearCommand extends CommandGroup {
 
   public DeliverGearCommand() {
     requires(Robot.driveSubsystem);
     requires(Robot.cameraSubsystem);
-
-//    anglePid = new PIDController(ANGLE_KP, ANGLE_KI, ANGLE_KD
-//            , new SimplePIDSource(Robot.driveSubsystem::getAngleToWall)
-//            , output -> commandedTurnRate = output);
-//    anglePid.setAbsoluteTolerance(ANGLE_TOLERANCE_DEGREES);
-//    anglePid.setSetpoint(0);
-
-    strafePid = new PIDController(STRAFE_KP, STRAFE_KI, STRAFE_KD,
-            new PIDSource() {
-              @Override
-              public void setPIDSourceType(PIDSourceType pidSource) {
-
-              }
-
-              @Override
-              public PIDSourceType getPIDSourceType() {
-                return null;
-              }
-
-              @Override
-              public double pidGet() {
-                Robot.cameraSubsystem.doVisionProcessingOnGearCamera();
-                if (Robot.cameraSubsystem.isGearTargetAcquired()) {
-                  return Robot.cameraSubsystem.getGearTargetPixelsFromCenter();
-                } else {
-                  return 0;
-                }
-              }
-            }
-            , this::strafePidOutput);
-    strafePid.setAbsoluteTolerance(STRAFE_TOLERANCE);
-    strafePid.setInputRange(-160, 160);
-    strafePid.setOutputRange(-1, 1);
-    strafePid.setSetpoint(0);
-    SmartDashboard.putData("strafepid", strafePid);
-
-//    approachPid = new PIDController(APPROACH_KP, APPROACH_KI, APPROACH_KD
-//            , new SimplePIDSource(Robot.driveSubsystem::getDistanceToWall)
-//            , output -> commandedApproachRate = output);
-//    approachPid.setAbsoluteTolerance(APPROACH_TOLERANCE);
-//    strafePid.setSetpoint(0.25); // todo: calibrate this distance
+    addParallel(new DeliverGearStrafeCommand());
   }
-
-  protected void strafePidOutput(double output) {
-    commandedStrafeRate = output;
-    SmartDashboard.putNumber("strafeoutput", output);
-  }
-
-  @Override
-  protected void execute() {
-    strafePid.enable();
-    SmartDashboard.putNumber("commandedStrafeRate", commandedStrafeRate);
-    //Robot.driveSubsystem.mecanumDrive(commandedStrafeRate, commandedApproachRate, commandedTurnRate, 0);
-    Robot.driveSubsystem.mecanumDrive(commandedStrafeRate, 0, 0, 0); // temporarily replaced approach and turn rate because we don't yet have ultrasonics
-  }
-
-  @Override
-  protected boolean isFinished() {
-    return strafePid.onTarget();
-    //return anglePid.onTarget() && strafePid.onTarget() && approachPid.onTarget();
-  }
-
-  private double offsetToLift() {
-    Robot.cameraSubsystem.doVisionProcessingOnGearCamera();
-    if (Robot.cameraSubsystem.isGearTargetAcquired()) {
-      return Robot.cameraSubsystem.getGearTargetPixelsFromCenter();
-    } else {
-      return 0;
-    }
-  }
+//  private static final double ANGLE_TOLERANCE_DEGREES = 2.0f;
+//  private static final double ANGLE_KP = 0.3f;
+//  private static final double ANGLE_KI = 0;
+//  private static final double ANGLE_KD = 0;
+//
+//  private static final double STRAFE_TOLERANCE = 0.5f;
+//  private static final double STRAFE_KP = 0.3f;
+//  private static final double STRAFE_KI = 0;
+//  private static final double STRAFE_KD = 0;
+//
+//  private static final double APPROACH_TOLERANCE = 0.5f;
+//  private static final double APPROACH_KP = 0.3f;
+//  private static final double APPROACH_KI = 0;
+//  private static final double APPROACH_KD = 0;
+//
+//  //private final PIDController anglePid;
+//  private final PIDController strafePid;
+//  //private final PIDController approachPid;
+//
+//  private double commandedTurnRate = 0;
+//  private double commandedStrafeRate = 0;
+//  private double commandedApproachRate = 0;
+//
+//  public DeliverGearCommand() {
+//    requires(Robot.driveSubsystem);
+//    requires(Robot.cameraSubsystem);
+//
+////    anglePid = new PIDController(ANGLE_KP, ANGLE_KI, ANGLE_KD
+////            , new SimplePIDSource(Robot.driveSubsystem::getAngleToWall)
+////            , output -> commandedTurnRate = output);
+////    anglePid.setAbsoluteTolerance(ANGLE_TOLERANCE_DEGREES);
+////    anglePid.setSetpoint(0);
+//
+//    strafePid = new PIDController(STRAFE_KP, STRAFE_KI, STRAFE_KD,
+//            new PIDSource() {
+//              @Override
+//              public void setPIDSourceType(PIDSourceType pidSource) {
+//
+//              }
+//
+//              @Override
+//              public PIDSourceType getPIDSourceType() {
+//                return null;
+//              }
+//
+//              @Override
+//              public double pidGet() {
+//                Robot.cameraSubsystem.doVisionProcessingOnGearCamera();
+//
+//                if (Robot.cameraSubsystem.isGearTargetAcquired()) {
+//                  return Robot.cameraSubsystem.getGearTargetPixelsFromCenter();
+//                } else {
+//                  return 0;
+//                }
+//              }
+//            }
+//            , this::strafePidOutput);
+//    strafePid.setAbsoluteTolerance(STRAFE_TOLERANCE);
+//    strafePid.setInputRange(-160, 160);
+//    strafePid.setOutputRange(-1, 1);
+//    strafePid.setSetpoint(0);
+//    SmartDashboard.putData("strafepid", strafePid);
+//
+////    approachPid = new PIDController(APPROACH_KP, APPROACH_KI, APPROACH_KD
+////            , new SimplePIDSource(Robot.driveSubsystem::getDistanceToWall)
+////            , output -> commandedApproachRate = output);
+////    approachPid.setAbsoluteTolerance(APPROACH_TOLERANCE);
+////    strafePid.setSetpoint(0.25); // todo: calibrate this distance
+//  }
+//
+//  protected void strafePidOutput(double output) {
+//    commandedStrafeRate = output;
+//    SmartDashboard.putNumber("strafeoutput", output);
+//  }
+//
+//  @Override
+//  protected void execute() {
+//    strafePid.enable();
+//    SmartDashboard.putNumber("commandedStrafeRate", commandedStrafeRate);
+//    //Robot.driveSubsystem.mecanumDrive(commandedStrafeRate, commandedApproachRate, commandedTurnRate, 0);
+//    Robot.driveSubsystem.mecanumDrive(commandedStrafeRate, 0, 0, 0); // temporarily replaced approach and turn rate because we don't yet have ultrasonics
+//  }
+//
+//  @Override
+//  protected boolean isFinished() {
+//    return strafePid.onTarget();
+//    //return anglePid.onTarget() && strafePid.onTarget() && approachPid.onTarget();
+//  }
+//
+//  private double offsetToLift() {
+//    Robot.cameraSubsystem.doVisionProcessingOnGearCamera();
+//    if (Robot.cameraSubsystem.isGearTargetAcquired()) {
+//      return Robot.cameraSubsystem.getGearTargetPixelsFromCenter();
+//    } else {
+//      return 0;
+//    }
+//  }
 
 }
